% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iadbget.raw}
\alias{iadbget.raw}
\title{Call the Data Catalog API}
\usage{
iadbget.raw(url)
}
\arguments{
\item{url}{A charcter string. A formatted url string}
}
\value{
A data frame
}
\description{
Helper function for the data catalog call
}
\note{
This call is seperate because the data catalog is actaully a different
 API and therefore has a different return structure.
}

