% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadbsearch.R
\name{iadbmsearch}
\alias{iadbmsearch}
\title{Get the metadata of the indicators}
\usage{
iadbmsearch(value = "ALL", lang = "EN")
}
\arguments{
\item{value}{Character string. 'ALL' or the IndicatorCode}

\item{lang}{Language. Default EN: English ES for Spanish}
}
\value{
Data frame with IndicatorCode, DataSetName, IndicatorName, IndicatorLongName, IndicatorShortDef,
IndicatorSynonym, IndicatorSource, TopicID, TopicName, SubTopicName, AggregationMethod, BasePeriod, Derivation Method,
OtherNote, Periodicity, PowerCode, Predefinedregion, PredefinedCountry, redefinedTopic, ReferencePerior, UOM, 
GeneralComment, LimitationException and SourceComment
}
\description{
Data frame with series code, name, description, and source for the IADB
}
\examples{
iadbmsearch()
iadbmsearch(value="SOC_050")
}
