% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadbstats.R
\name{iadbstats.countries}
\alias{iadbstats.countries}
\title{Get the metadata of the countries available in IADB}
\usage{
iadbstats.countries()
}
\value{
Data frame with CountryCode, ALPHA2Code, WB2Code, CountryShortName, CountryLongName, CountryIncomeGroup, OtherRegionalGroup, CountrySpecialNotes
}
\description{
Data frame of countries. 1 row per country
}
\examples{
iadbstats.countries()
}
