% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadbstats.R
\name{iadbstats}
\alias{iadbstats}
\title{Pull the data for some indicators and countries.}
\usage{
iadbstats(country = "ALL", frequency = "year", indicatorcode = "ALL")
}
\arguments{
\item{country}{Character string. ISO3 code of the countries. E.g COL for Colombia or 'ALL' for all the countries}

\item{frequency}{Character string. Frequency of the indicators Year for yearly.}

\item{indicatorcode}{codes of the indicators separated by comma. 'All' for all the indicators}
}
\value{
Data frame with  CountryCode, CountryTableName, IndicatorCode, IndicatorName, TopicName, SubTopicName, Year, Quarter,
AggregationLevel, AggregatedValue, UOM
}
\description{
Data frame with the data of the specified indicators and countries
All countries and All indicators cannot be requested
When All indicators data is requested, data can be requested for a maximum of 4 countries
When All country data is requested, data can be requested for a maximum of 10 indicators (see iadbstats.list)
}
\examples{
iadbstats(country="COL",frequency="year",indicatorcode="SOC_050")
iadbstats(country="ARG,COL",frequency="year",indicatorcode="SOC_050")
}
