% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iadburls}
\alias{iadburls}
\title{Internal function to get the urls of the API}
\usage{
iadburls()
}
\value{
List with 2 entries. The first one with the base_url and the second one with the utils_url#'
}
\description{
List with 2 entries. The first one with the base_url and the second one with the utils_url#'
}
